pyproject.toml
[project]
authors = [{name = "Benedict Ruppel", email = "benedict.ruppel@tum.de"}]
name = "large-scale-vit-slam"
requires-python = ">= 3.10"
version = "0.0.1"
readme = "README.md"
dependencies = [
    "numpy<2",
    "Pillow",
    "huggingface_hub",
    "einops",
    "safetensors",
    "opencv-python",
    "lightning==2.5.1",
    "pandas",
    "torchmetrics",
    "hydra-core",
    "omegaconf",
    "onnxruntime",
    "viser==0.2.23",
    "matplotlib",
    "scipy",
    "requests",
    "wcmatch",
    "gradio",
]

# Using setuptools as the build backend
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# setuptools configuration
[tool.setuptools.packages.find]
where = ["."]
include = ["aligned_vggt*"]